Broke VersionSpec and VersionOrder into a separate module
